feat: 상품 관리자 수정 페이지 옵션 반영

+++ src/main/java/gift/option/service/OptionServiceImpl.java
@@ -72,6 +72,10 @@ public class OptionServiceImpl implements OptionService{
             throw new OptionNotFoundException();
         }
 
+        if (optionRepository.findAllByProductId(productId).size() <= 1) {
+            throw new MinimumQuantityViolationException();
+        }
+
         optionRepository.deleteById(optionId);
     }
